requirements.txt
Flask
Flask-Session
Werkzeug
cs50
requests
python-dotenv
